Started on TvP. Lots of work on combat and more passive play involving siege tanks and defensive positions which is apparently necessary in this matchup.
Added strat_tvp_opening, which currently tries to get siege tanks with siege mode out as quickly as possible, accompanied by a few marines to hold any early aggression. If the opponent takes a fast expansion, then we do too with a bunker.
strat_tank is supposed to be the followup, which is very basic right now; it wants to make siege tanks and vultures and attacks when it has 12 siege tanks.