strat_tvt_opening: open 11 rax 12 refinery, set combat::defence_is_scared.
Explicitly open 11 rax 12 refinery (same as in other matchups), to make sure the barracks does not get delayed for some reason (necessary to hold cheese).
If we are defending aggression (and we don't have the upper hand), set combat::defence_is_scared, which will send our early marines to the mineral line.