strat_tvp_opening: try to deal with dt rushes and don't wall in.
We build a missile turret in the hopes that a dt will run into it.
The wall-in was causing more harm than good at the moment, I would like to re-enable it at some point when I improve the unit interactions with it.
We now go spider mines before siege mode, since they're nice to have to kill dts and siege tanks without siege mode are excellent at dealing with early dragoons.
We're also more careful about when we expand