combat: some small but impactful changes regarding workers, tanks, defensive_choke and scanning.
Into the mid/late-game, try to not use scvs for fighting or repairing, since it usually results in losing too many.
Let multiple vultures lay down defensive spider mines simultaneously.
Introduced some unused code. yey.
For small groups, if we outnumber them, don't try to play defensively with siege tanks. Should help defend drops.
Commented out the part that made us always want to fight with siege tanks. It's possible that I still want to do this, but only for the actual siege tanks. Requires more testing.
We now scan like crazy at targets that are in the fog. Probably too much. The intent is primarily to give vision for siege tanks. (Need to write code for using flying units for this purpose.)
Wee now calculate the paths from the main base to all our other bases and choose the point closest to the nearest opponent base as my_closest_base (which probably needs a name-change now), thus that is essentially the point at which we want to defend. This helps move our units out on the map as we take more bases, hopefully in a better position to protect all of them.