strat_tvt_opening: we now open (usually 3 factory) vultures and go into double expand.
Not sure if this is a viable opening, the double expand might be too greedy, and the natural is taken a little late, but it seems entirely safe against every other terran bot that I've tested, mostly because vultures with spider mines are pretty good.
The 3 factories usually sit around idle for a while while we double expand, but the idea is that just in case the opponent masses vultures, we delay the expanding and instead match his vultures.