combat: lots of additions related to zerg units. Some experimental stuff.
Added support for defilers (dark swarm, plague, consume) and  queens (just spawn broodlings yet).
Kind of lazily fixed some issues with units not attacking spider mines. Might need better solution.
Added support for lurkers, though it's buggy and inefficient.
Experimenting with multiplying the eval score with 1.5/0.5 based on whether are are currently engaged in a fight or not, in the hopes that we will commit more when attacking. Unusure how it affects terran as opposed to zerg.
Cleaned up the code a bit, and fixed some minor issues.