combat: mostly stuff related to zerg, some experimental.
Let can_transfer_to return true for places that have very little ground threat (such as overlords, so that scouting overlords won't keep us from mining).
Added no_scout_around, which can be set by strategies to prevent units from running around and scouting (when attacking).
Trying to tweak the inside_group logic, such that it does not trigger for units that are running away from us, nor units that we should ignore unless we were assigned to kill them.
For zerglings, we will pull back those who are about to die if we are already winning the fight. Frequently they will soon run back in and die. :)
For defensive groups, we will now assume a defensive concave-like formation and hopefully maintain that formation while fighting, in order to get better surface area than the opponent. Effective in zergling vs zergling fights, less so otherwise, but it should end up attacking normally if it is better to do so.
Also wrote some code to consider whether we should actually run when we would like to do so. If we would take too much damage, we just fight instead, in the hope of being more cost-effective than if we just ran.
Some other tweaks and minor stuff