build: let hatcheries be built on creep.
Previously it would build them off creep, which spread them around. This can be bad when we get rushed, since we might build in-base hatcheries that are outside range of static defence. The only thing that concerns me is that this results in much more compact base layout, and ultralisks could get stuck (since they are 38x32).